Restructure map folders
