+++ __tests__/maps/diocesesMap.test.tsx
@@ -1,18 +1,19 @@
 jest.mock("react-leaflet", () => ({
-    MapContainer: jest.fn(() => null),
-    TileLayer: jest.fn(() => null),
-    Polygon: jest.fn(() => null),
-  }));
+  MapContainer: jest.fn(() => null),
+  TileLayer: jest.fn(() => null),
+  Polygon: jest.fn(() => null),
+}));
 
 import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
 
-import { displayBaseMap, displayDiocesesLayer } from "@/pages/components/DiocesesMap";
+import {
+  displayBaseMap,
+  displayDiocesesLayer,
+} from "@/pages/components/maps/DiocesesMap";
 import { TileLayer, Polygon, MapContainer } from "react-leaflet";
 
-
-
-import DiocesesMap from "@/pages/components/DiocesesMap";
+import DiocesesMap from "@/pages/components/maps/DiocesesMap";
 
 describe("Map section", () => {
   it("should render a map section", () => {
@@ -31,35 +32,33 @@ describe("Map section", () => {
     expect(heading).toHaveTextContent("Dioceses Map");
   });
 
-
-  function baseMapWrapper(){
-    return displayBaseMap()
+  function baseMapWrapper() {
+    return displayBaseMap();
   }
 
   it("should display the open street base map with copyright notice", async () => {
     const expectedCopyright =
       '<a href="https://www.maptiler.com/copyright/" target="_blank">© MapTiler</a> <a href="https://www.openstreetmap.org/copyright" target="_blank">© OpenStreetMap contributors</a>';
     const expectedBaseMapUrl =
-      'https://api.maptiler.com/maps/basic-v2/{z}/{x}/{y}.png?key=KEkSXTZni1ygdkq0Q875';
+      "https://api.maptiler.com/maps/basic-v2/{z}/{x}/{y}.png?key=KEkSXTZni1ygdkq0Q875";
 
     render(baseMapWrapper());
     expect(TileLayer).toHaveBeenCalledWith(
-        expect.objectContaining({
-            attribution: expectedCopyright,
-            url: expectedBaseMapUrl,
-          }),
-          undefined //undefined because we're not building the full DOM here
+      expect.objectContaining({
+        attribution: expectedCopyright,
+        url: expectedBaseMapUrl,
+      }),
+      undefined //undefined because we're not building the full DOM here
     );
   });
 
-  function diocesesLayerWrapper(){
-    return displayDiocesesLayer()
+  function diocesesLayerWrapper() {
+    return displayDiocesesLayer();
   }
 
-  it("should create polygons for all 21 dioceses", async () =>{
-    render(diocesesLayerWrapper())
+  it("should create polygons for all 21 dioceses", async () => {
+    render(diocesesLayerWrapper());
 
-    expect(Polygon).toHaveBeenCalledTimes(21)
-
-  })
+    expect(Polygon).toHaveBeenCalledTimes(21);
+  });
 });
